Break extensions into their own files
